feat: conform Search Documents comps to Lookup
